Add tests for response caching
