+++ client/src/main/java/net/spals/oembed4j/client/registry/DefaultOEmbedRegistry.java
@@ -10,6 +10,7 @@ import net.spals.oembed4j.model.OEmbedProvider;
 
 import java.io.File;
 import java.io.IOException;
+import java.io.InputStream;
 import java.net.URI;
 import java.util.Collections;
 import java.util.List;
